+++ packages/pigeon/lib/java_generator.dart
@@ -503,13 +503,20 @@ void generateJava(JavaOptions options, Root root, StringSink sink) {
     indent.writeln('import java.util.HashMap;');
   }
 
+  String _camelToSnake(String camelCase) {
+    final RegExp regex = RegExp('([a-z])([A-Z]+)');
+    return camelCase
+        .replaceAllMapped(regex, (Match m) => '${m[1]}_${m[2]}')
+        .toUpperCase();
+  }
+
   void writeEnum(Enum anEnum) {
     indent.write('public enum ${anEnum.name} ');
     indent.scoped('{', '}', () {
       int index = 0;
       for (final String member in anEnum.members) {
         indent.writeln(
-            '$member($index)${index == anEnum.members.length - 1 ? ';' : ','}');
+            '${_camelToSnake(member)}($index)${index == anEnum.members.length - 1 ? ';' : ','}');
         index++;
       }
       indent.writeln('');