+++ packages/pigeon/lib/pigeon_lib.dart
@@ -329,11 +329,6 @@ class ParseResults {
   final Map<String, Object>? pigeonOptions;
 }
 
-String _posixify(String input) {
-  final path.Context context = path.Context(style: path.Style.posix);
-  return context.fromUri(path.toUri(path.absolute(input)));
-}
-
 Iterable<String> _lineReader(String path) sync* {
   final String contents = File(path).readAsStringSync();
   const LineSplitter lineSplitter = LineSplitter();
@@ -424,17 +419,14 @@ class DartTestGenerator implements Generator {
 
   @override
   void generate(StringSink sink, PigeonOptions options, Root root) {
-    final String mainPath = path.context.relative(
-      _posixify(options.dartOut!),
-      from: _posixify(path.dirname(options.dartTestOut!)),
-    );
     final DartOptions dartOptionsWithHeader = _dartOptionsWithCopyrightHeader(
         options.dartOptions, options.copyrightHeader);
     generateTestDart(
       dartOptionsWithHeader,
       root,
       sink,
-      mainPath,
+      dartOutPath: options.dartOut!,
+      testOutPath: options.dartTestOut!,
     );
   }
 
@@ -886,7 +878,7 @@ class _RootBuilder extends dart_ast_visitor.RecursiveAstVisitor<Object?> {
     _storeCurrentApi();
     _storeCurrentClass();
 
-    if (node.isAbstract) {
+    if (node.abstractKeyword != null) {
       if (_hasMetadata(node.metadata, 'HostApi')) {
         final dart_ast.Annotation hostApi = node.metadata.firstWhere(
             (dart_ast.Annotation element) => element.name.name == 'HostApi');
@@ -907,20 +899,20 @@ class _RootBuilder extends dart_ast_visitor.RecursiveAstVisitor<Object?> {
           }
         }
         _currentApi = Api(
-          name: node.name.name,
+          name: node.name2.lexeme,
           location: ApiLocation.host,
           methods: <Method>[],
           dartHostTestHandler: dartHostTestHandler,
         );
       } else if (_hasMetadata(node.metadata, 'FlutterApi')) {
         _currentApi = Api(
-          name: node.name.name,
+          name: node.name2.lexeme,
           location: ApiLocation.flutter,
           methods: <Method>[],
         );
       }
     } else {
-      _currentClass = Class(name: node.name.name, fields: <NamedType>[]);
+      _currentClass = Class(name: node.name2.lexeme, fields: <NamedType>[]);
     }
 
     node.visitChildren(this);
@@ -940,7 +932,7 @@ class _RootBuilder extends dart_ast_visitor.RecursiveAstVisitor<Object?> {
               baseName: argTypeBaseName,
               isNullable: isNullable,
               typeArguments: argTypeArguments),
-          name: parameter.identifier?.name ?? '',
+          name: parameter.name?.lexeme ?? '',
           offset: parameter.offset);
     } else {
       return NamedType(
@@ -1003,7 +995,7 @@ class _RootBuilder extends dart_ast_visitor.RecursiveAstVisitor<Object?> {
       final dart_ast.SimpleIdentifier returnTypeIdentifier =
           getFirstChildOfType<dart_ast.SimpleIdentifier>(returnType)!;
       _currentApi!.methods.add(Method(
-          name: node.name.name,
+          name: node.name2.lexeme,
           returnType: TypeDeclaration(
               baseName: returnTypeIdentifier.name,
               typeArguments: typeAnnotationsToTypeArguments(
@@ -1017,7 +1009,7 @@ class _RootBuilder extends dart_ast_visitor.RecursiveAstVisitor<Object?> {
     } else if (_currentClass != null) {
       _errors.add(Error(
           message:
-              'Methods aren\'t supported in Pigeon data classes ("${node.name.name}").',
+              'Methods aren\'t supported in Pigeon data classes ("${node.name2.lexeme}").',
           lineNumber: _calculateLineNumber(source, node.offset)));
     }
     node.visitChildren(this);
@@ -1027,9 +1019,9 @@ class _RootBuilder extends dart_ast_visitor.RecursiveAstVisitor<Object?> {
   @override
   Object? visitEnumDeclaration(dart_ast.EnumDeclaration node) {
     _enums.add(Enum(
-        name: node.name.name,
+        name: node.name2.lexeme,
         members: node.constants
-            .map((dart_ast.EnumConstantDeclaration e) => e.name.name)
+            .map((dart_ast.EnumConstantDeclaration e) => e.name2.lexeme)
             .toList()));
     node.visitChildren(this);
     return null;
@@ -1075,7 +1067,7 @@ class _RootBuilder extends dart_ast_visitor.RecursiveAstVisitor<Object?> {
                   baseName: type.name.name,
                   isNullable: type.question != null,
                   typeArguments: typeAnnotationsToTypeArguments(typeArguments)),
-              name: node.fields.variables[0].name.name,
+              name: node.fields.variables[0].name2.lexeme,
               offset: node.offset));
         }
       } else {