+++ packages/pigeon/test/java_generator_test.dart
@@ -41,7 +41,8 @@ void main() {
       name: 'Foobar',
       members: <String>[
         'one',
-        'two',
+        'twoThreeFour',
+        'remoteDB',
       ],
     );
     final Root root = Root(
@@ -54,8 +55,9 @@ void main() {
     generateJava(javaOptions, root, sink);
     final String code = sink.toString();
     expect(code, contains('public enum Foobar'));
-    expect(code, contains('    one(0),'));
-    expect(code, contains('    two(1);'));
+    expect(code, contains('    ONE(0),'));
+    expect(code, contains('    TWO_THREE_FOUR(1),'));
+    expect(code, contains('    REMOTE_DB(2);'));
     expect(code, contains('private int index;'));
     expect(code, contains('private Foobar(final int index) {'));
     expect(code, contains('      this.index = index;'));
@@ -583,7 +585,8 @@ void main() {
       name: 'Enum1',
       members: <String>[
         'one',
-        'two',
+        'twoThreeFour',
+        'remoteDB',
       ],
     );
     final Class klass = Class(
@@ -608,8 +611,9 @@ void main() {
     generateJava(javaOptions, root, sink);
     final String code = sink.toString();
     expect(code, contains('public enum Enum1'));
-    expect(code, contains('    one(0),'));
-    expect(code, contains('    two(1);'));
+    expect(code, contains('    ONE(0),'));
+    expect(code, contains('    TWO_THREE_FOUR(1),'));
+    expect(code, contains('    REMOTE_DB(2);'));
     expect(code, contains('private int index;'));
     expect(code, contains('private Enum1(final int index) {'));
     expect(code, contains('      this.index = index;'));